MWPW-158484: Round 1 Blog Performance Improvements (#73)
* Performance improvement parity with bacom
