+++ head.html
@@ -1,5 +1,26 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
 <script src="/blog/scripts/fallback.js" nomodule></script>
+<script>
+  const libs = (() => {
+    const { hostname, search } = window.location;
+    if (!['.hlx.', '.stage.', 'local'].some((i) => hostname.includes(i))) return '/libs';
+    const branch = new URLSearchParams(search).get('milolibs') || 'main';
+    if (branch === 'local') return 'http://localhost:6456/libs';
+    return branch.includes('--') ? `https://${branch}.hlx.live/libs` : `https://${branch}--milo--adobecom.hlx.live/libs`;
+  })();
+
+  const miloStyles = document.createElement('link');
+  const miloUtils = document.createElement('link');
+
+  miloStyles.setAttribute('as', 'style');
+  miloStyles.setAttribute('href', `${libs}/styles/styles.css`);
+  miloUtils.setAttribute('as', 'script');
+  miloUtils.setAttribute('crossorigin', 'true');
+  miloUtils.setAttribute('href', `${libs}/utils/utils.js`);
+
+  [miloStyles, miloUtils].forEach((tag) => tag.setAttribute('rel', 'preload'));
+  document.head.append(miloStyles, miloUtils);
+</script>
 <script src="/blog/scripts/scripts.js" type="module"></script>
 <style>body { display: none; }</style>
 <link rel="icon" href="data:,">
